modified mesh2d & sectionGenerator to support python3.x
modified mesh2d.py to support python3.x, used 2to3 convert

modified line 133, convert the len() to int, since in python3, the divide will result float by default

+++ sectionGenerator.py
@@ -130,7 +130,7 @@ def RHS(d, b, t, r_out, n_r):
         points.append((x_out, y_out))
         points.append((x_in, y_in))
 
-    for i in range(len(points) / 2):
+    for i in range(int(len(points) / 2)):
         if i != len(points) / 2 - 1:
             facets.append((i * 2, i * 2 + 2))
             facets.append((i * 2 + 1, i * 2 + 3))
